eMMC: Enable eMMC at boot
After reworking the init script the eMMC stopped to be detected by
Linux. It turns out that the eMMC controller needs to be initialized by
U-Boot.

Introduce a PREBOOT command to initialize the eMMC and cleanup the code
path to reduce the excessive spam and to remove also the code related to
the aml-specific way of managing the eMMC partitions.

Signed-off-by: Carlo Caione <carlo@endlessm.com>